添加了动画素材

+++ momoka/include/core/tools/Gui.h
@@ -7,7 +7,7 @@ class Gui {
 public:
 	void DrawDebugText(std::string text);
 	void DrawHeroHp(float currentValue, float maxValue, int posX, int posY);
-	void DrawEnemyHp(int currentValue, int maxValue, int posX, int posY);
+	void DrawEnemyHp(float currentValue, float maxValue, int posX, int posY);
 
 	Gui();
 };
